Implement high score into the App
Every game now is connected to the store to update scores when a game
is finished. Redux handles sorting and reducing the scores, so that
the scores that are displayed on GameView on the front end are in
descending order, only show the top ten scores, and only show one score
from each player regardless of how many scores that player may have on
a particular game.